// src/app/shared/models/mathInterfaces.ts
export interface MathInterfacesLevels {
  firstView: string;
  secondView: string;
  level: number;
  isActive: boolean;
}

export interface MathInterfacesMenuItem {
  name: string;
  id: number;
  urlPath: string;
}
export interface MathInterfacesAnswerStatistic {
  questionAnswer: string;
  spentTime: any;
  timeStamp: number;
}